+++ client/main/js/views/cart.js
@@ -8,26 +8,29 @@ app.CartView = Backbone.View.extend({
 		this.$el.html(template);
 	},
 	events: {
-		"click .button_cancel": "cancel",
+		"click .button_cancel": "cancelSeries",
 		"click .button_cancel_next_week": "cancelNextWeek"
 	},
-	cancel: function(event) {
+	cancelSeries: function(event) {
 		//
-		$.ajax({
-			url: '/cart',
-			type: 'delete',
-			data: {
-				tutor: app.objectBuffer.tutor,
-				customer: "sunj3", //the user who is currently logged in
-				course: app.objectBuffer.course,
-				time: app.objectBuffer.time
-			},
-			success: function() {
-				header_view.showCart();	
-			}
-		});
-		
-		console.log("send emails to both client and tutor");
+		var response = confirm("Are you sure you want to cancel this appointment?");
+		if (response) {
+			$.ajax({
+				url: '/cart',
+				type: 'delete',
+				data: {
+					tutor: app.objectBuffer.tutor,
+					customer: "sunj3", //the user who is currently logged in
+					course: app.objectBuffer.course,
+					time: app.objectBuffer.time
+				},
+				success: function() {
+					header_view.showCart();	
+				}
+			});
+			
+			console.log("send emails to both client and tutor");
+		}
 		
 	},
 	cancelNextWeek: function(event) {
